refactor GetPassphrase to shared pkg
