+++ cmd/appliance/backup/api_test.go
@@ -186,68 +186,3 @@ func TestBackupAPICommand(t *testing.T) {
 		t.Fatalf("Expected output\n%s\ngot\n%s\n", want, got)
 	}
 }
-
-func Test_getPassPhrase(t *testing.T) {
-	type args struct {
-		stdIn     io.Reader
-		canPrompt bool
-		hasStdin  bool
-	}
-	tests := []struct {
-		name     string
-		args     args
-		askStubs func(*prompt.AskStubber)
-		want     string
-		wantErr  bool
-	}{
-		{
-			name: "with stdin",
-			args: args{
-				stdIn:     bytes.NewBuffer([]byte("hunter2\n")),
-				canPrompt: false,
-				hasStdin:  true,
-			},
-			want:    "hunter2",
-			wantErr: false,
-		},
-		{
-			name: "with prompt",
-			args: args{
-				canPrompt: true,
-				hasStdin:  false,
-			},
-			want:    "secret",
-			wantErr: false,
-			askStubs: func(s *prompt.AskStubber) {
-				s.StubPrompt("The passphrase to encrypt the appliance backups when the Backup API is used:").AnswerWith("secret")
-				s.StubPrompt("Confirm your passphrase:").AnswerWith("secret")
-			},
-		},
-		{
-			name: "no stdin no prompt",
-			args: args{
-				canPrompt: false,
-				hasStdin:  false,
-			},
-			want:    "",
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			stubber, teardown := prompt.InitAskStubber(t)
-			defer teardown()
-			if tt.askStubs != nil {
-				tt.askStubs(stubber)
-			}
-			got, err := getPassPhrase(tt.args.stdIn, tt.args.canPrompt, tt.args.hasStdin)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("getPassPhrase() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("getPassPhrase() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
