+++ pkg/prompt/password.go
@@ -2,8 +2,12 @@ package prompt
 
 import (
 	"errors"
+	"fmt"
+	"io"
+	"strings"
 
 	"github.com/AlecAivazis/survey/v2"
+	"github.com/appgate/sdpctl/pkg/cmdutil"
 )
 
 func PasswordConfirmation(message string) (string, error) {
@@ -24,4 +28,20 @@ func PasswordConfirmation(message string) (string, error) {
 		return firstAnswer, errors.New("The passphrase did not match")
 	}
 	return firstAnswer, nil
+}
+
+// GetPassphrase check stdin if we have anything, and use that as passphrase
+// otherwise, if we can prompt, Prompt user input
+func GetPassphrase(stdIn io.Reader, canPrompt, hasStdin bool, message string) (string, error) {
+	if hasStdin {
+		buf, err := io.ReadAll(stdIn)
+		if err != nil {
+			return "", fmt.Errorf("could not read input from stdin %s", err)
+		}
+		return strings.TrimSuffix(string(buf), "\n"), nil
+	}
+	if !canPrompt {
+		return "", cmdutil.ErrMissingTTY
+	}
+	return PasswordConfirmation(message)
 }
